refactor(plugins): separate inventory reads from lifecycle mutations
