+++ src/cli/plugins-install-command.ts
@@ -5,7 +5,7 @@ import { reportClawHubPluginInstallTelemetry } from "../infra/clawhub-packages.j
 import { parseClawHubPluginSpec } from "../infra/clawhub-spec.js";
 import { formatErrorMessage } from "../infra/errors.js";
 import { CLAWHUB_INSTALL_ERROR_CODE } from "../plugins/clawhub.js";
-import { installManagedPluginSource } from "../plugins/management-service.js";
+import { installManagedPluginSource } from "../plugins/management-install.js";
 import { withPluginLifecycleLease } from "../plugins/plugin-lifecycle-lease.js";
 import { tracePluginLifecyclePhaseAsync } from "../plugins/plugin-lifecycle-trace.js";
 import { defaultRuntime } from "../runtime.js";
