+++ src/gateway/server-methods/plugins.test.ts
@@ -20,9 +20,12 @@ const searchMock = vi.hoisted(() => vi.fn());
 
 vi.mock("../../plugins/management-service.js", () => ({
   inspectManagedPlugin: (...args: unknown[]) => managementMocks.inspect(...args),
-  installManagedPlugin: (...args: unknown[]) => managementMocks.install(...args),
   listManagedPlugins: (...args: unknown[]) => managementMocks.list(...args),
   refreshManagedPluginMetadata: (...args: unknown[]) => managementMocks.refreshMetadata(...args),
+}));
+
+vi.mock("../../plugins/management-mutations.js", () => ({
+  installManagedPlugin: (...args: unknown[]) => managementMocks.install(...args),
   setManagedPluginEnabled: (...args: unknown[]) => managementMocks.setEnabled(...args),
   uninstallManagedPlugin: (...args: unknown[]) => managementMocks.uninstall(...args),
 }));
@@ -31,7 +34,9 @@ vi.mock("../../plugins/catalog-search.js", () => ({
   searchInstallablePluginPackages: (...args: unknown[]) => searchMock(...args),
 }));
 
-const { pluginsHandlers } = await import("./plugins.js");
+const { pluginsHandlers: pluginReadHandlers } = await import("./plugins.js");
+const { pluginMutationHandlers } = await import("./plugins-mutations.js");
+const pluginsHandlers = { ...pluginReadHandlers, ...pluginMutationHandlers };
 
 async function callHandler(
   method: string,
