+++ src/plugins/management-service.ts
@@ -1,40 +1,14 @@
-// Structured plugin catalog and lifecycle operations shared by Gateway-facing surfaces.
+// Reads plugin inventory and inspection metadata without loading install orchestration.
 import { redactSensitiveUrlLikeString } from "@openclaw/net-policy/redact-sensitive-url";
-import { asSafeIntegerInRange } from "@openclaw/normalization-core/number-coercion";
 import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
-import { uniqueStrings } from "@openclaw/normalization-core/string-normalization";
 import type {
   PluginInspectSource,
   PluginsInspectResult,
 } from "../../packages/gateway-protocol/src/schema/plugins.js";
-import { MANIFEST_KEY } from "../compat/legacy-names.js";
-import { collectChangedPaths } from "../config/config-change-paths.js";
-import {
-  assertConfigWriteAllowedInCurrentMode,
-  readConfigFileSnapshotForWrite,
-  replaceConfigFile,
-} from "../config/config.js";
 import { resolveConfigWidePluginMetadataSnapshot } from "../config/io.plugin-metadata.js";
 import { resolveIsNixMode } from "../config/paths.js";
-import { ensurePluginAllowlisted } from "../config/plugins-allowlist.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
-import type { PluginInstallRecord } from "../config/types.plugins.js";
-import { parseClawHubPluginSpec } from "../infra/clawhub-spec.js";
-import { formatErrorMessage } from "../infra/errors.js";
-import { buildNpmResolutionFields, type NpmSpecResolution } from "../infra/install-source-utils.js";
-import { parseRegistryNpmSpec } from "../infra/npm-registry-spec.js";
-import { normalizeUpdateChannel, resolveRegistryUpdateChannel } from "../infra/update-channels.js";
-import type { RuntimeEnv } from "../runtime.js";
-import { VERSION } from "../version.js";
-import { installBundledPluginSource } from "./bundled-install.js";
-import type { BundledPluginSource } from "./bundled-sources.js";
-import {
-  prepareManagedPluginArtifactConsentHandler,
-  resolvePendingPluginCapabilityReview,
-  resolvePluginCapabilityConsent,
-  type PluginCapabilityConsentAcknowledgment,
-  type PluginCapabilityConsentHandler,
-} from "./capability-consent.js";
+import { resolvePendingPluginCapabilityReview } from "./capability-consent.js";
 import {
   buildPluginCapabilitySummary,
   computeDeclaredSurfaceHash,
@@ -44,94 +18,45 @@ import {
   resolvePluginInstallRecordTrust,
   resolvePluginPackageDeclaredSurface,
 } from "./capability-summary.js";
-import { CLAWHUB_INSTALL_ERROR_CODE, isUnavailableClawHubTarget } from "./clawhub-error-codes.js";
-import {
-  buildClawHubPluginInstallRecordFields,
-  type ClawHubPluginInstallRecordFields,
-} from "./clawhub-install-records.js";
-import { installPluginFromClawHub } from "./clawhub.js";
 import {
   appendPluginControlPlaneWorkspaceDiagnostic,
   resolvePluginControlPlaneWorkspace,
 } from "./control-plane-workspace.js";
 import { getProcessGatewayPluginMetadataSnapshot } from "./current-plugin-metadata-state.js";
-import { enableExplicitlySelectedPluginInConfig } from "./enable.js";
-import { installPluginFromGitSpec } from "./git-install.js";
-import {
-  installWithSourceFallback,
-  type PluginInstallSource,
-  resolveClawHubInstallSpecsForUpdateChannel,
-  resolveNpmInstallSpecsForUpdateChannel,
-} from "./install-channel-specs.js";
-import { resolveDefaultPluginExtensionsDir } from "./install-paths.js";
-import {
-  resolveInstallConfigMutationPreflights,
-  selectInstallMutationWriteOptions,
-  persistPluginInstall,
-  type ConfigSnapshotForInstallPersist,
-} from "./install-persistence.js";
-import { commitPluginInstallRecordsWithConfig } from "./install-record-commit.js";
-import type { InstallSafetyOverrides } from "./install-security-scan.js";
-import type { InstallPolicyWarningDetails } from "./install-security-scan.types.js";
-import {
-  requestDeferredPluginInstall,
-  resolvePluginInstallTransaction,
-  type PluginInstallTransaction,
-} from "./install-transaction.js";
-import {
-  isUnavailableNpmTarget,
-  PLUGIN_INSTALL_ERROR_CODE,
-  type PluginInstallArtifactConsentRequest,
-  type PluginInstallLogger,
-} from "./install-types.js";
-import {
-  installPluginFromNpmPackArchive,
-  installPluginFromNpmSpec,
-  installPluginFromPath,
-} from "./install.js";
-import {
-  loadInstalledPluginIndexInstallRecords,
-  removePluginInstallRecordFromRecords,
-  withPluginInstallRecords,
-  withoutPluginInstallRecords,
-} from "./installed-plugin-index-records.js";
-import { createInstalledPluginIndexScopeLookup } from "./installed-plugin-index-scope-lookup.js";
 import {
   createInstalledPluginEnabledPredicate,
   isInstalledPluginEnabled,
 } from "./installed-plugin-index.js";
 import { createInstalledPluginOwnershipResolver } from "./installed-plugin-package-ownership.js";
+import {
+  type ManagedPluginCatalogEntry,
+  type ManagedPluginCatalog,
+  type OfficialCatalogResult,
+  getManagedPluginCache,
+  withManagedPluginCache,
+  prepareCatalogEntry,
+  prepareCatalogEntries,
+  loadOfficialCatalog,
+  normalizeKinds,
+  normalizeCatalogMetadata,
+  normalizeFeaturedAt,
+  derivePluginCategory,
+  firstPluginError,
+  compareCatalogEntries,
+  resolveInstalledPluginPresentation,
+  resolveInstalledHostedOfficialEntry,
+  resolveOfficialEntryById,
+} from "./management-catalog.js";
 import { ManagedPluginLifecycleError } from "./management-lifecycle-error.js";
-import type { PluginManifestRecord } from "./manifest-registry.js";
 import type { PluginDiagnostic } from "./manifest-types.js";
-import { installPluginFromMarketplace } from "./marketplace.js";
 import {
-  resolveTrustedOfficialClawHubPackageName,
-  resolveTrustedSourceLinkedOfficialClawHubSpec,
-  resolveTrustedSourceLinkedOfficialNpmSpec,
-} from "./official-external-install-records.js";
-import {
-  getOfficialExternalPluginCatalogEntryForPackage,
   getOfficialExternalPluginCatalogManifest,
   listOfficialExternalPluginCatalogEntries,
-  loadConfiguredHostedOfficialExternalPluginCatalogEntries,
   resolveOfficialExternalPluginId,
-  resolveOfficialExternalPluginInstall,
-  resolveOfficialExternalPluginInstallSources,
   resolveOfficialExternalPluginLabel,
-  type HostedOfficialExternalPluginCatalogLoadResult,
-  type OfficialExternalPluginCatalogEntry,
 } from "./official-external-plugin-catalog.js";
 import { tracksPluginDependencyStatus } from "./official-external-plugin-repair-hints.js";
-import {
-  createPluginCache,
-  getPluginCache,
-  getPluginMetadataSnapshotCache,
-  getProcessPluginCache,
-  getScopedPluginCache,
-  withPluginCache,
-} from "./plugin-cache.js";
-import { withPluginLifecycleLease } from "./plugin-lifecycle-lease.js";
+import { createPluginCache, getProcessPluginCache, withPluginCache } from "./plugin-cache.js";
 import {
   loadPluginMetadataSnapshot,
   resolvePluginMetadataSnapshot,
@@ -139,202 +64,12 @@ import {
 } from "./plugin-metadata-snapshot.js";
 import { resolveManifestProviderAuthChoices } from "./provider-auth-choices.js";
 import { listRecommendedToolInstalls } from "./recommended-tool-installs.js";
-import { refreshPluginRegistryAfterConfigMutation } from "./registry-refresh.js";
-import { applySlotSelectionForPlugin } from "./slot-selection.js";
 import {
   buildPluginDependencyStatus,
   projectPluginDependencyHealth,
 } from "./status-dependencies-core.js";
-import { setPluginEnabledInConfig } from "./toggle-config.js";
-import { collectClawPluginUninstallWarnings } from "./uninstall-claw-references.js";
-import {
-  prepareConfigForDisabledPluginSet,
-  recordPluginPackageUninstallPlan,
-} from "./uninstall-package-plan.js";
-import {
-  applyPluginUninstallDirectoryRemoval,
-  formatUninstallActionLabels,
-  planPluginUninstall,
-  pluginUninstallTargetExists,
-} from "./uninstall.js";
-
-type ManagedPluginCatalogEntry = {
-  id: string;
-  name: string;
-  packageName?: string;
-  description?: string;
-  version?: string;
-  kind?: string[];
-  origin?: string;
-  installed: boolean;
-  enabled: boolean;
-  state: "enabled" | "disabled" | "not-installed" | "error";
-  featured?: boolean;
-  featuredAt?: number;
-  order?: number;
-  hasIcon?: boolean;
-  install?: { source: "clawhub"; packageName: string } | { source: "official"; pluginId: string };
-  error?: string;
-  category?: string;
-  removable?: boolean;
-};
-
-type ManagedPluginCatalog = {
-  plugins: ManagedPluginCatalogEntry[];
-  diagnostics: unknown[];
-  mutationAllowed: boolean;
-};
 
 export type ManagedPluginInspection = PluginsInspectResult;
-
-type ManagedPluginInstallRequest =
-  | {
-      source: "clawhub";
-      packageName: string;
-      version?: string;
-      acknowledgeInstallPolicyWarning?: true;
-      acknowledgeCapabilities?: PluginCapabilityConsentAcknowledgment;
-    }
-  | {
-      source: "official";
-      pluginId: string;
-      acknowledgeInstallPolicyWarning?: true;
-      acknowledgeCapabilities?: PluginCapabilityConsentAcknowledgment;
-    };
-
-export type ManagedPluginSourceInstallRequest =
-  | {
-      source: "local";
-      path: string;
-      /** Stable source provenance when installing an owner-verified temporary copy. */
-      recordPath?: string;
-      recordSource: "archive" | "path";
-      mode: "install" | "update";
-      link?: boolean;
-      bundledOrigin?: true;
-      successMessage?: string;
-    }
-  | {
-      source: "npm-pack";
-      archivePath: string;
-      mode: "install" | "update";
-    }
-  | { source: "git"; spec: string; mode: "install" | "update" }
-  | {
-      source: "marketplace";
-      marketplace: string;
-      plugin: string;
-      mode: "install" | "update";
-    }
-  | {
-      source: "clawhub";
-      spec: string;
-      /** Spec recorded for the install; keeps user intent when `spec` is channel-resolved. */
-      recordSpec?: string;
-      mode?: "install" | "update";
-      expectedPluginId?: string;
-      expectedIntegrity?: string;
-      /** Host-validated official catalog provenance for release-cohort resolution. */
-      trustedSourceLinkedOfficialInstall?: true;
-      confirmInstall?: NonNullable<
-        Parameters<typeof installPluginFromClawHub>[0]["confirmInstall"]
-      >;
-    }
-  | {
-      source: "bundled";
-      rawSpec: string;
-      bundledSource: BundledPluginSource;
-      warning?: string;
-    }
-  | {
-      source: "official";
-      spec: string;
-      installSources?: PluginInstallSource[];
-      expectedPluginId?: string;
-      /** Spec recorded for the install; keeps user intent when `spec` is channel-resolved. */
-      recordSpec?: string;
-      pluginId: string;
-      expectedIntegrity?: string;
-      mode: "install" | "update";
-      pin?: boolean;
-    }
-  | {
-      source: "npm";
-      spec: string;
-      /** Spec recorded for the install; keeps user intent when `spec` is channel-resolved. */
-      recordSpec?: string;
-      mode: "install" | "update";
-      pin?: boolean;
-      expectedPluginId?: string;
-      expectedIntegrity?: string;
-      trustedSourceLinkedOfficialInstall?: boolean;
-      allowBundledFallback?: boolean;
-    };
-
-type ManagedPluginSourceInstallResult =
-  | {
-      ok: true;
-      pluginId: string;
-      config: OpenClawConfig;
-      warnings?: string[];
-      targetDir?: string;
-      version?: string;
-      npmResolution?: NpmSpecResolution;
-      clawhub?: ClawHubPluginInstallRecordFields;
-    }
-  | {
-      ok: false;
-      error: string;
-      code?: string;
-      version?: string;
-      warning?: string;
-      installPolicyWarning?: InstallPolicyWarningDetails;
-      installSource?: PluginInstallSource;
-    };
-
-type SourceInstallerResult =
-  | {
-      ok: false;
-      error: string;
-      code?: string;
-      version?: string;
-      warning?: string;
-      installPolicyWarning?: InstallPolicyWarningDetails;
-    }
-  | {
-      ok: true;
-      pluginId: string;
-      targetDir: string;
-      version?: string;
-      npmResolution?: NpmSpecResolution;
-    };
-
-type OfficialCatalogResult = Pick<HostedOfficialExternalPluginCatalogLoadResult, "entries"> & {
-  error?: string;
-  hostedFeaturedAuthoritative?: boolean;
-};
-
-function getManagedPluginCache(metadata?: PluginMetadataSnapshot) {
-  if (metadata) {
-    return getPluginMetadataSnapshotCache(metadata);
-  }
-  const scoped = getScopedPluginCache();
-  if (scoped?.kind === "operation") {
-    return scoped;
-  }
-  const candidate = getProcessPluginCache().desiredMetadata;
-  if (candidate && candidate.boot === getProcessGatewayPluginMetadataSnapshot()) {
-    return candidate.cache;
-  }
-  return getPluginCache();
-}
-
-function withManagedPluginCache<
-  TParams extends { config: OpenClawConfig; metadata?: PluginMetadataSnapshot },
-  TResult,
->(run: (params: TParams) => Promise<TResult>): (params: TParams) => Promise<TResult> {
-  return (params) => withPluginCache(getManagedPluginCache(params.metadata), () => run(params));
-}
 
 function resolveManagedPluginDiagnostics(
   snapshot: PluginMetadataSnapshot,
@@ -369,373 +104,6 @@ function resolveManagedPluginDiagnostics(
   return diagnostics;
 }
 
-/** Clear the process-stable hosted catalog snapshot after an explicit owner reload. */
-export function clearManagedPluginOfficialCatalogCache(): void {
-  getManagedPluginCache().officialCatalog = undefined;
-}
-
-function mergeCatalogMetadata(
-  hosted: OfficialExternalPluginCatalogEntry,
-  bundled: OfficialExternalPluginCatalogEntry,
-  options: { hostedFeaturedAuthoritative: boolean },
-): OfficialExternalPluginCatalogEntry {
-  const hostedManifest = getOfficialExternalPluginCatalogManifest(hosted);
-  const bundledManifest = getOfficialExternalPluginCatalogManifest(bundled);
-  const bundledCatalog = bundledManifest?.catalog;
-  const bundledPlugin = bundledManifest?.plugin;
-  const bundledName = normalizeOptionalString(bundled.name);
-  const bundledDescription = normalizeOptionalString(bundled.description);
-  const bundledKind = normalizeOptionalString(bundled.kind);
-  const bundledSource = normalizeOptionalString(bundled.source);
-  const hostedFeatured = typeof hosted.featured === "boolean" ? hosted.featured : false;
-  const mergedCatalog =
-    bundledCatalog ||
-    hostedManifest?.catalog ||
-    (options.hostedFeaturedAuthoritative && hostedFeatured)
-      ? {
-          ...hostedManifest?.catalog,
-          ...bundledCatalog,
-          ...(options.hostedFeaturedAuthoritative ? { featured: hostedFeatured } : {}),
-        }
-      : undefined;
-  if (!mergedCatalog && !bundledPlugin) {
-    return hosted;
-  }
-  return {
-    ...hosted,
-    ...(!normalizeOptionalString(hosted.name) && bundledName ? { name: bundledName } : {}),
-    ...(!normalizeOptionalString(hosted.description) && bundledDescription
-      ? { description: bundledDescription }
-      : {}),
-    ...(!normalizeOptionalString(hosted.kind) && bundledKind ? { kind: bundledKind } : {}),
-    ...(!normalizeOptionalString(hosted.source) && bundledSource ? { source: bundledSource } : {}),
-    [MANIFEST_KEY]: {
-      ...hostedManifest,
-      ...(bundledPlugin ? { plugin: { ...hostedManifest?.plugin, ...bundledPlugin } } : {}),
-      ...(mergedCatalog ? { catalog: mergedCatalog } : {}),
-    },
-  };
-}
-
-function prepareCatalogEntry(entry: OfficialExternalPluginCatalogEntry) {
-  const install = resolveOfficialExternalPluginInstall(entry);
-  const sources = resolveOfficialExternalPluginInstallSources(entry, { resolvedInstall: install });
-  const clawhubSpec = sources.find((source) => source.source === "clawhub")?.spec;
-  const npmSpec = sources.find((source) => source.source === "npm")?.spec;
-  return {
-    entry,
-    install,
-    selectedSource: sources[0],
-    clawhub: clawhubSpec ? parseClawHubPluginSpec(clawhubSpec) : undefined,
-    npmPackage: npmSpec ? parseRegistryNpmSpec(npmSpec)?.name : undefined,
-  };
-}
-
-type PreparedCatalogEntry = ReturnType<typeof prepareCatalogEntry>;
-
-function prepareCatalogEntries(entries: readonly OfficialExternalPluginCatalogEntry[]) {
-  let prepared: PreparedCatalogEntry[] | undefined;
-  // Unknown local installs never need package identities from the official catalog.
-  // Resolve each collection only when provenance admits a lookup, then reuse its facts.
-  return () => (prepared ??= entries.map(prepareCatalogEntry));
-}
-
-/**
- * Overlay local runtime identity and ordering after an exact package/source match.
- * Hosted curation wins; bundled Featured state survives only in fallback mode.
- */
-function overlayBundledOfficialPluginCatalogMetadata(
-  entries: readonly OfficialExternalPluginCatalogEntry[],
-  bundledEntries: readonly OfficialExternalPluginCatalogEntry[] = listOfficialExternalPluginCatalogEntries(),
-  options: { hostedFeaturedAuthoritative: boolean } = {
-    hostedFeaturedAuthoritative: false,
-  },
-): OfficialExternalPluginCatalogEntry[] {
-  const bundledFacts = entries.length > 0 ? bundledEntries.map(prepareCatalogEntry) : [];
-  return entries.map((entry) => {
-    const { clawhub, npmPackage } = prepareCatalogEntry(entry);
-    const matches = bundledFacts.filter(
-      (bundled) =>
-        (clawhub && bundled.clawhub?.name === clawhub.name) ||
-        (npmPackage && bundled.npmPackage === npmPackage),
-    );
-    const bundled = matches.length === 1 ? matches[0]?.entry : undefined;
-    if (bundled) {
-      return mergeCatalogMetadata(entry, bundled, options);
-    }
-    if (!options.hostedFeaturedAuthoritative) {
-      return entry;
-    }
-    const hostedManifest = getOfficialExternalPluginCatalogManifest(entry);
-    if (entry.featured !== true && !hostedManifest?.catalog) {
-      return entry;
-    }
-    return {
-      ...entry,
-      [MANIFEST_KEY]: {
-        ...hostedManifest,
-        catalog: {
-          ...hostedManifest?.catalog,
-          featured: entry.featured === true,
-        },
-      },
-    };
-  });
-}
-
-async function loadOfficialCatalog(): Promise<OfficialCatalogResult> {
-  const cache = getManagedPluginCache();
-  if (!cache.officialCatalog) {
-    const promise = Promise.resolve().then(() =>
-      loadConfiguredHostedOfficialExternalPluginCatalogEntries(),
-    );
-    cache.officialCatalog = promise;
-    void promise.catch(() => {
-      if (cache.officialCatalog === promise) {
-        cache.officialCatalog = undefined;
-      }
-    });
-  }
-  const result = await cache.officialCatalog;
-  const hostedFeaturedAuthoritative =
-    result.source === "hosted" || result.source === "hosted-snapshot";
-  return {
-    entries: overlayBundledOfficialPluginCatalogMetadata(result.entries, undefined, {
-      hostedFeaturedAuthoritative,
-    }),
-    hostedFeaturedAuthoritative,
-    ...("error" in result ? { error: result.error } : {}),
-  };
-}
-
-function normalizeKinds(kind: string | readonly string[] | undefined): string[] | undefined {
-  const values = (typeof kind === "string" ? [kind] : (kind ?? []))
-    .map((value) => value.trim())
-    .filter(Boolean);
-  return values.length > 0 ? [...new Set(values)] : undefined;
-}
-
-function normalizeCatalogMetadata(
-  value: unknown,
-): { featured?: boolean; order?: number } | undefined {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return undefined;
-  }
-  const record = value as Record<string, unknown>;
-  const featured = typeof record.featured === "boolean" ? record.featured : undefined;
-  const order =
-    typeof record.order === "number" && Number.isFinite(record.order) ? record.order : undefined;
-  return featured === undefined && order === undefined
-    ? undefined
-    : {
-        ...(featured !== undefined ? { featured } : {}),
-        ...(order !== undefined ? { order } : {}),
-      };
-}
-
-function normalizeFeaturedAt(value: unknown): number | undefined {
-  return asSafeIntegerInRange(value, { min: 0 });
-}
-
-/** Coarse manifest-derived grouping so catalog UIs can shelve a large inventory. */
-function derivePluginCategory(manifest: PluginManifestRecord | undefined): string | undefined {
-  if (!manifest) {
-    return undefined;
-  }
-  if (manifest.channels.length > 0 || Object.keys(manifest.channelConfigs ?? {}).length > 0) {
-    return "channel";
-  }
-  const mediaProvider =
-    Object.keys(manifest.imageGenerationProviderMetadata ?? {}).length > 0 ||
-    Object.keys(manifest.videoGenerationProviderMetadata ?? {}).length > 0 ||
-    Object.keys(manifest.musicGenerationProviderMetadata ?? {}).length > 0 ||
-    Object.keys(manifest.mediaUnderstandingProviderMetadata ?? {}).length > 0;
-  if (
-    manifest.providers.length > 0 ||
-    manifest.providerEndpoints?.length ||
-    manifest.modelCatalog ||
-    mediaProvider
-  ) {
-    return "provider";
-  }
-  const kinds = normalizeKinds(manifest.kind);
-  if (kinds?.includes("memory")) {
-    return "memory";
-  }
-  if (kinds?.includes("context-engine")) {
-    return "context-engine";
-  }
-  if (
-    manifest.contracts?.tools?.length ||
-    Object.keys(manifest.toolMetadata ?? {}).length > 0 ||
-    manifest.skills.length > 0
-  ) {
-    return "tool";
-  }
-  return undefined;
-}
-
-function firstPluginError(
-  diagnostics: readonly PluginDiagnostic[],
-  pluginId: string,
-): string | undefined {
-  return diagnostics.find(
-    (diagnostic) => diagnostic.level === "error" && diagnostic.pluginId === pluginId,
-  )?.message;
-}
-
-function compareCatalogEntries(
-  left: ManagedPluginCatalogEntry,
-  right: ManagedPluginCatalogEntry,
-): number {
-  const featured = Number(Boolean(right.featured)) - Number(Boolean(left.featured));
-  if (featured !== 0) {
-    return featured;
-  }
-  if (left.featured && right.featured) {
-    const leftFeaturedAt = left.featuredAt;
-    const rightFeaturedAt = right.featuredAt;
-    if (leftFeaturedAt !== undefined || rightFeaturedAt !== undefined) {
-      if (leftFeaturedAt === undefined) {
-        return 1;
-      }
-      if (rightFeaturedAt === undefined) {
-        return -1;
-      }
-      if (leftFeaturedAt !== rightFeaturedAt) {
-        return rightFeaturedAt - leftFeaturedAt;
-      }
-    }
-  }
-  const order = (left.order ?? Number.MAX_SAFE_INTEGER) - (right.order ?? Number.MAX_SAFE_INTEGER);
-  return order !== 0 ? order : left.name.localeCompare(right.name);
-}
-
-function resolveInstalledOfficialCatalogEntry(params: {
-  entries: ReturnType<typeof prepareCatalogEntries>;
-  packageName?: string;
-  source: "clawhub" | "npm";
-}): PreparedCatalogEntry | undefined {
-  if (!params.packageName) {
-    return undefined;
-  }
-  const matches = params
-    .entries()
-    .filter(
-      ({ clawhub, npmPackage }) =>
-        (params.source === "clawhub" ? clawhub?.name : npmPackage) === params.packageName,
-    );
-  return matches.length === 1 ? matches[0] : undefined;
-}
-
-type PluginIndexRecord = PluginMetadataSnapshot["index"]["plugins"][number];
-
-function resolveInstalledPluginPresentation(params: {
-  record: PluginIndexRecord;
-  manifest?: PluginManifestRecord;
-  officialEntry?: OfficialExternalPluginCatalogEntry;
-  hostedListingAuthoritative: boolean;
-}): Pick<ManagedPluginCatalogEntry, "name" | "description" | "version"> {
-  const { record, manifest, officialEntry, hostedListingAuthoritative } = params;
-  // Registry names may be backfilled with npm specifiers, which are not display labels.
-  const manifestName = manifest?.name !== record.packageName ? manifest?.name : undefined;
-  const localName = manifestName ?? manifest?.channelCatalogMeta?.label ?? record.pluginId;
-  const localDescription =
-    manifest?.description ?? manifest?.channelCatalogMeta?.blurb ?? manifest?.packageDescription;
-  const name =
-    (hostedListingAuthoritative ? normalizeOptionalString(officialEntry?.title) : undefined) ??
-    localName;
-  const description =
-    (hostedListingAuthoritative
-      ? normalizeOptionalString(officialEntry?.description)
-      : undefined) ?? localDescription;
-  const version = record.packageVersion ?? manifest?.version;
-  return {
-    name,
-    ...(description ? { description } : {}),
-    ...(version ? { version } : {}),
-  };
-}
-
-function resolveInstalledHostedOfficialEntry(params: {
-  record: PluginIndexRecord;
-  installOwner?: string;
-  installRecord?: PluginInstallRecord;
-  officialEntries: ReturnType<typeof prepareCatalogEntries>;
-  bundledOfficialEntries: ReturnType<typeof prepareCatalogEntries>;
-}): {
-  entry?: OfficialExternalPluginCatalogEntry;
-  clawhubPackage?: string;
-} {
-  const identityPluginId = params.installOwner ?? params.record.pluginId;
-  const trustedOfficialClawHubSpec = params.installRecord
-    ? resolveTrustedSourceLinkedOfficialClawHubSpec({
-        pluginId: identityPluginId,
-        record: params.installRecord,
-      })
-    : undefined;
-  const trustedOfficialNpmSpec = params.installRecord
-    ? resolveTrustedSourceLinkedOfficialNpmSpec({
-        pluginId: identityPluginId,
-        record: params.installRecord,
-      })
-    : undefined;
-  const sourceLinkedOfficialClawHubPackage = trustedOfficialClawHubSpec
-    ? parseClawHubPluginSpec(trustedOfficialClawHubSpec)?.name
-    : undefined;
-  const currentOfficialClawHubPackage = params.installRecord
-    ? resolveTrustedOfficialClawHubPackageName(params.installRecord)
-    : undefined;
-  const trustedOfficialNpmPackage = trustedOfficialNpmSpec
-    ? parseRegistryNpmSpec(trustedOfficialNpmSpec)?.name
-    : undefined;
-  const bundledPublishedEntry =
-    params.record.origin === "bundled"
-      ? resolveInstalledOfficialCatalogEntry({
-          entries: params.bundledOfficialEntries,
-          packageName: params.record.packageName,
-          source: "npm",
-        })
-      : undefined;
-  const installedOfficialIdentity = sourceLinkedOfficialClawHubPackage
-    ? { source: "clawhub" as const, packageName: sourceLinkedOfficialClawHubPackage }
-    : trustedOfficialNpmPackage
-      ? { source: "npm" as const, packageName: trustedOfficialNpmPackage }
-      : currentOfficialClawHubPackage &&
-          (!params.record.packageName ||
-            params.record.packageName === currentOfficialClawHubPackage)
-        ? { source: "clawhub" as const, packageName: currentOfficialClawHubPackage }
-        : bundledPublishedEntry && params.record.packageName
-          ? { source: "npm" as const, packageName: params.record.packageName }
-          : undefined;
-  const hasInstalledOfficialProvenance = Boolean(
-    installedOfficialIdentity &&
-    (!params.record.packageName ||
-      params.record.packageName === installedOfficialIdentity.packageName),
-  );
-  const bundledOfficialEntry =
-    bundledPublishedEntry ??
-    resolveInstalledOfficialCatalogEntry({
-      entries: params.bundledOfficialEntries,
-      packageName: hasInstalledOfficialProvenance
-        ? installedOfficialIdentity?.packageName
-        : undefined,
-      source: installedOfficialIdentity?.source ?? "clawhub",
-    });
-  const hostedPackageName =
-    installedOfficialIdentity?.source === "npm"
-      ? bundledOfficialEntry?.clawhub?.name
-      : installedOfficialIdentity?.packageName;
-  return {
-    entry: resolveInstalledOfficialCatalogEntry({
-      entries: params.officialEntries,
-      packageName: hasInstalledOfficialProvenance ? hostedPackageName : undefined,
-      source: "clawhub",
-    })?.entry,
-    clawhubPackage: hasInstalledOfficialProvenance ? hostedPackageName : undefined,
-  };
-}
-
 export type ManagedPluginIconSource = { kind: "file"; path: string; rootPath: string };
 
 function resolvePluginIconSource(params: {
@@ -767,7 +135,7 @@ function resolveManagedPluginMetadata(config: OpenClawConfig, env: NodeJS.Proces
     : resolvePluginMetadataSnapshot(resolveManagedPluginMetadataParams(config, env));
 }
 
-function loadFreshManagedPluginMetadata(config: OpenClawConfig, env: NodeJS.ProcessEnv) {
+export function loadFreshManagedPluginMetadata(config: OpenClawConfig, env: NodeJS.ProcessEnv) {
   // Gateway actions must cover every workspace shown in its management inventory.
   return getProcessGatewayPluginMetadataSnapshot()
     ? resolveConfigWidePluginMetadataSnapshot({ config, env, allowCurrent: false })
@@ -1212,1024 +580,3 @@ export const inspectManagedPlugin = withManagedPluginCache(
     };
   },
 );
-
-function assertValidConfigSnapshot(
-  prepared: Awaited<ReturnType<typeof readConfigFileSnapshotForWrite>>,
-): ConfigSnapshotForInstallPersist {
-  const { snapshot, writeOptions } = prepared;
-  if (!snapshot.valid) {
-    throw new ManagedPluginLifecycleError(
-      "Config invalid; run `openclaw doctor --fix` before managing plugins.",
-    );
-  }
-  const mutationWriteOptions = selectInstallMutationWriteOptions(writeOptions);
-  const { pluginMutation } = resolveInstallConfigMutationPreflights({
-    parsed: (snapshot.parsed ?? {}) as Record<string, unknown>,
-    snapshotPath: snapshot.path,
-    writeOptions: mutationWriteOptions,
-  });
-  if (pluginMutation.mode === "blocked") {
-    throw new ManagedPluginLifecycleError(pluginMutation.reason);
-  }
-  return {
-    config: snapshot.sourceConfig,
-    baseHash: snapshot.hash,
-    writeOptions: mutationWriteOptions,
-  };
-}
-
-async function readPluginMutationSnapshot(
-  env: NodeJS.ProcessEnv,
-): Promise<ConfigSnapshotForInstallPersist> {
-  try {
-    assertConfigWriteAllowedInCurrentMode({ env });
-  } catch (error) {
-    throw new ManagedPluginLifecycleError(formatErrorMessage(error), { cause: error });
-  }
-  return assertValidConfigSnapshot(await readConfigFileSnapshotForWrite());
-}
-
-function createSilentRuntime(): RuntimeEnv {
-  return {
-    log: () => undefined,
-    error: () => undefined,
-    exit: (code) => {
-      throw new ManagedPluginLifecycleError(`plugin lifecycle exited with code ${code}`);
-    },
-  };
-}
-
-function createInstallLogger(warnings: string[]) {
-  return {
-    info: () => undefined,
-    warn: (message: string) => warnings.push(message),
-  };
-}
-
-function resolveOfficialEntryById(
-  entries: readonly OfficialExternalPluginCatalogEntry[],
-  pluginId: string,
-): OfficialExternalPluginCatalogEntry | undefined {
-  return entries.find((entry) => resolveOfficialExternalPluginId(entry) === pluginId);
-}
-
-/** Explicitly declared runtime id, ignoring the entry-id fallback used for display. */
-function resolveDeclaredOfficialPluginId(
-  entry: OfficialExternalPluginCatalogEntry,
-): string | undefined {
-  const manifest = getOfficialExternalPluginCatalogManifest(entry);
-  return (
-    normalizeOptionalString(manifest?.plugin?.id) ??
-    normalizeOptionalString(manifest?.channel?.id) ??
-    normalizeOptionalString(manifest?.providers?.[0]?.id)
-  );
-}
-
-function resolveOfficialEntryByClawHubPackage(
-  entries: readonly OfficialExternalPluginCatalogEntry[],
-  packageName: string,
-): OfficialExternalPluginCatalogEntry | undefined {
-  // Bundled identities remain the local trust anchor when a hosted feed omits
-  // its ClawHub candidate; hosted install/version metadata is never copied back.
-  return [...listOfficialExternalPluginCatalogEntries(), ...entries].find((entry) => {
-    return resolveOfficialExternalPluginInstallSources(entry).some(
-      (source) =>
-        source.source === "clawhub" && parseClawHubPluginSpec(source.spec)?.name === packageName,
-    );
-  });
-}
-
-function resolveHostedOfficialEntryByClawHubPackage(
-  entries: readonly OfficialExternalPluginCatalogEntry[],
-  packageName: string,
-): OfficialExternalPluginCatalogEntry | undefined {
-  return entries.find((entry) => {
-    return resolveOfficialExternalPluginInstallSources(entry).some(
-      (source) =>
-        source.source === "clawhub" && parseClawHubPluginSpec(source.spec)?.name === packageName,
-    );
-  });
-}
-
-function buildClawHubSpec(packageName: string, version?: string): string {
-  const parsed = parseClawHubPluginSpec(`clawhub:${packageName}`);
-  if (!parsed || parsed.version) {
-    throw new ManagedPluginLifecycleError(`invalid ClawHub package name: ${packageName}`);
-  }
-  return `clawhub:${packageName}${version ? `@${version}` : ""}`;
-}
-
-function throwInstallFailure(result: {
-  error: string;
-  code?: string;
-  version?: string;
-  warning?: string;
-  installPolicyWarning?: InstallPolicyWarningDetails;
-}): never {
-  const unavailable =
-    !result.code ||
-    result.code === CLAWHUB_INSTALL_ERROR_CODE.ARTIFACT_UNAVAILABLE ||
-    result.code === CLAWHUB_INSTALL_ERROR_CODE.ARTIFACT_DOWNLOAD_UNAVAILABLE ||
-    result.code === CLAWHUB_INSTALL_ERROR_CODE.CLAWHUB_SECURITY_UNAVAILABLE;
-  throw new ManagedPluginLifecycleError(result.error, {
-    kind: unavailable ? "unavailable" : "invalid-request",
-    code: result.code,
-    version: result.version,
-    warning: result.warning,
-    installPolicyWarning: result.installPolicyWarning,
-    cause: result,
-  });
-}
-
-async function persistManagedSourceInstall(params: {
-  snapshot: ConfigSnapshotForInstallPersist;
-  pluginId: string;
-  install: PluginInstallRecord;
-  transaction?: PluginInstallTransaction;
-  invalidateRuntimeCache?: boolean;
-  runtime?: RuntimeEnv;
-  successMessage?: string;
-  beforePersistentApply?: () => void;
-  beforePersistentEffect?: () => void | Promise<void>;
-}): Promise<{ config: OpenClawConfig; warnings: string[] }> {
-  const warnings: string[] = [];
-  let committed = false;
-  try {
-    const config = await persistPluginInstall({
-      snapshot: params.snapshot,
-      pluginId: params.pluginId,
-      install: params.install,
-      invalidateRuntimeCache: params.invalidateRuntimeCache,
-      runtime: params.runtime,
-      persistenceLogger: { warn: (message) => warnings.push(message) },
-      beforePersistentApply: params.beforePersistentApply,
-      beforePersistentEffect: params.beforePersistentEffect,
-      // Only the persistence owner can distinguish rejection from a late refresh failure.
-      onCommitted: () => {
-        committed = true;
-      },
-      ...(params.successMessage ? { successMessage: params.successMessage } : {}),
-    });
-    return { config, warnings };
-  } catch (error) {
-    if (!committed) {
-      try {
-        await params.transaction?.rollback();
-      } catch (rollbackError) {
-        // Both errors are retained; the install failure remains the primary cause.
-        const aggregate = new AggregateError(
-          [error, rollbackError],
-          "Plugin install failed and payload rollback failed",
-        );
-        aggregate.cause = error;
-        throw aggregate;
-      }
-    }
-    throw error;
-  } finally {
-    if (committed) {
-      await params.transaction?.commit().catch(() => {
-        const warning = "Plugin install committed, but backup cleanup failed. Restart is required.";
-        warnings.push(warning);
-        params.runtime?.log(warning);
-      });
-    }
-  }
-}
-
-/**
- * Official plugin installs target the release stream the gateway is running,
- * the same target `openclaw doctor --fix` and `openclaw plugins update`
- * already resolve. Resolving here keeps every managed install path — CLI,
- * chat command, and any future caller — on one answer instead of letting the
- * registry default land a plugin the gateway then reports as drifted.
- *
- * Beta and extended-stable resolve here. Version-bound stable tracks key off a
- * per-plugin `versionBoundToOpenClaw` descriptor that a managed install request
- * does not carry, and answering for them from this boundary would pin plugins
- * the policy never opted in.
- */
-function resolveOfficialManagedInstallSpec(params: {
-  request: Extract<ManagedPluginSourceInstallRequest, { source: "official" | "npm" | "clawhub" }>;
-  config: OpenClawConfig;
-}): string | null {
-  const { request } = params;
-  const trustedSourceLinkedOfficialInstall =
-    request.source !== "official" && request.trustedSourceLinkedOfficialInstall === true;
-  if (request.source === "npm" && !trustedSourceLinkedOfficialInstall) {
-    return null;
-  }
-  // An integrity pin identifies one exact artifact, so it outranks the channel.
-  if (request.expectedIntegrity) {
-    return null;
-  }
-  const packageName =
-    request.source === "clawhub"
-      ? parseClawHubPluginSpec(request.spec)?.name
-      : parseRegistryNpmSpec(request.spec)?.name;
-  if (
-    !packageName ||
-    (request.source !== "official" &&
-      !trustedSourceLinkedOfficialInstall &&
-      !getOfficialExternalPluginCatalogEntryForPackage(packageName))
-  ) {
-    return null;
-  }
-  const updateChannel = resolveRegistryUpdateChannel({
-    configChannel: normalizeUpdateChannel(params.config.update?.channel),
-    currentVersion: VERSION,
-  });
-  if (updateChannel !== "beta" && updateChannel !== "extended-stable") {
-    return null;
-  }
-  const specs =
-    request.source === "clawhub"
-      ? resolveClawHubInstallSpecsForUpdateChannel({
-          spec: request.spec,
-          updateChannel,
-          officialPackageName: packageName,
-          coreVersion: VERSION,
-        })
-      : resolveNpmInstallSpecsForUpdateChannel({
-          spec: request.spec,
-          updateChannel,
-          officialPackageName: packageName,
-          coreVersion: VERSION,
-        });
-  return specs.installSpec === request.spec ? null : specs.installSpec;
-}
-
-type ManagedPluginSourceInstallParams = {
-  request: ManagedPluginSourceInstallRequest;
-  snapshot: ConfigSnapshotForInstallPersist;
-  env?: NodeJS.ProcessEnv;
-  logger?: PluginInstallLogger & { terminalLinks?: boolean };
-  safetyOverrides?: InstallSafetyOverrides;
-  runtime?: RuntimeEnv;
-  invalidateRuntimeCache?: boolean;
-  acknowledgeCapabilities?: PluginCapabilityConsentAcknowledgment;
-  onCapabilityConsent?: PluginCapabilityConsentHandler;
-  beforePersistentApply?: () => void;
-  /** Revalidate the initiating owner after artifact review and before durable activation. */
-  beforePersistentEffect?: () => void | Promise<void>;
-};
-
-/**
- * Installs official plugins from the release stream the gateway runs. When that
- * stream has no published artifact the install reports it instead of widening
- * back to the registry default: widening would resolve `latest` and land exactly
- * the cross-release plugin this boundary exists to prevent, and a fresh install
- * has nothing to preserve, so failing with the reason costs the operator only a
- * retry with an explicit version.
- */
-export async function installManagedPluginSource(
-  params: ManagedPluginSourceInstallParams,
-): Promise<ManagedPluginSourceInstallResult> {
-  return await withPluginLifecycleLease({ env: params.env }, async (lease) => {
-    const assertOwned = lease.assertOwned.bind(lease);
-    const ownedParams = {
-      ...params,
-      beforePersistentApply: () => {
-        params.beforePersistentApply?.();
-        assertOwned();
-      },
-    };
-    return await installManagedPluginSourceUnderLease(ownedParams, assertOwned);
-  });
-}
-
-async function installManagedPluginSourceUnderLease(
-  params: ManagedPluginSourceInstallParams,
-  assertOwned: () => void,
-): Promise<ManagedPluginSourceInstallResult> {
-  const { request } = params;
-  if (request.source === "official" && request.installSources) {
-    const { attempt: installAttempt, source: installedSource } = await installWithSourceFallback({
-      sources: request.pin
-        ? request.installSources.filter((source) => source.source === "npm")
-        : request.installSources,
-      install: async (source) =>
-        await installManagedPluginSource({
-          ...params,
-          request: {
-            source: source.source,
-            spec: source.spec,
-            mode: request.mode,
-            expectedPluginId: request.expectedPluginId,
-            trustedSourceLinkedOfficialInstall: true,
-            ...(source.expectedIntegrity ? { expectedIntegrity: source.expectedIntegrity } : {}),
-            ...(source.source === "npm" && request.pin ? { pin: true } : {}),
-          },
-        }),
-      result: (attempt) => attempt,
-      onFallback: (message) => params.logger?.warn?.(message),
-    });
-    return installAttempt.ok
-      ? installAttempt
-      : { ...installAttempt, installSource: installedSource };
-  }
-  if (request.source !== "official" && request.source !== "npm" && request.source !== "clawhub") {
-    return await installResolvedManagedPluginSource(params, assertOwned);
-  }
-  const installSpec = resolveOfficialManagedInstallSpec({
-    request,
-    config: params.snapshot.config,
-  });
-  if (!installSpec) {
-    return await installResolvedManagedPluginSource(params, assertOwned);
-  }
-  const result = await installResolvedManagedPluginSource(
-    {
-      ...params,
-      request: { ...request, spec: installSpec, recordSpec: request.recordSpec ?? request.spec },
-    },
-    assertOwned,
-  );
-  if (result.ok) {
-    return result;
-  }
-  const isUnavailableTarget =
-    request.source === "clawhub"
-      ? isUnavailableClawHubTarget(result)
-      : isUnavailableNpmTarget(result);
-  if (!isUnavailableTarget) {
-    return result;
-  }
-  return {
-    ...result,
-    code: PLUGIN_INSTALL_ERROR_CODE.RELEASE_COHORT_UNAVAILABLE,
-    error: `No ${installSpec} release is published for this gateway. Installing ${request.spec} would resolve a build from another release; pass an explicit version to install one anyway.`,
-  };
-}
-
-/** Execute one resolved plugin source through the shared install-and-persist pipeline. */
-async function installResolvedManagedPluginSource(
-  params: ManagedPluginSourceInstallParams,
-  assertOwned: () => void,
-): Promise<ManagedPluginSourceInstallResult> {
-  const { request } = params;
-  const env = params.env ?? process.env;
-  const extensionsDir = resolveDefaultPluginExtensionsDir(env);
-  if (request.source === "bundled") {
-    const result = await installBundledPluginSource({
-      ...params,
-      rawSpec: request.rawSpec,
-      bundledSource: request.bundledSource,
-      warning: request.warning,
-    });
-    return {
-      ok: true,
-      ...result,
-      config: params.snapshot.config,
-    };
-  }
-
-  const consentExemptSource = request.source === "local" && request.bundledOrigin === true;
-  const source =
-    request.source === "local"
-      ? request.recordSource
-      : request.source === "npm-pack" || request.source === "official"
-        ? "npm"
-        : request.source;
-  const capabilityConsent = consentExemptSource
-    ? undefined
-    : await prepareManagedPluginArtifactConsentHandler({
-        config: params.snapshot.config,
-        env,
-        source,
-        ...(request.source === "marketplace"
-          ? { spec: `${request.plugin}@${request.marketplace}` }
-          : "spec" in request
-            ? { spec: request.spec }
-            : {}),
-        ...("expectedIntegrity" in request && request.expectedIntegrity
-          ? { expectedIntegrity: request.expectedIntegrity }
-          : {}),
-        acknowledgeCapabilities: params.acknowledgeCapabilities,
-        onCapabilityConsent: params.onCapabilityConsent,
-      });
-
-  const common = requestDeferredPluginInstall(
-    {
-      ...params.safetyOverrides,
-      config: params.snapshot.config,
-      extensionsDir,
-      logger: params.logger,
-      beforePersistentApply: params.beforePersistentApply,
-      ...(capabilityConsent || params.beforePersistentEffect
-        ? {
-            onBeforePluginArtifactCommit: async (artifact: PluginInstallArtifactConsentRequest) => {
-              await capabilityConsent?.onBeforePluginArtifactCommit(artifact);
-              await params.beforePersistentEffect?.();
-            },
-          }
-        : {}),
-    },
-    undefined,
-    assertOwned,
-  );
-  const complete = async <T extends SourceInstallerResult>(
-    installResult: Promise<T>,
-    completed: {
-      install: (result: Extract<T, { ok: true }>) => PluginInstallRecord;
-      expectedPluginId?: string;
-      snapshot?: ConfigSnapshotForInstallPersist;
-      successMessage?: string;
-    },
-  ): Promise<ManagedPluginSourceInstallResult> => {
-    const result = await installResult;
-    if (!result.ok) {
-      return result;
-    }
-    const installed = result as Extract<T, { ok: true }> & {
-      pluginId: string;
-      targetDir: string;
-    };
-    // Linking skips the installer's staging transaction but still grants durable authority.
-    if (request.source === "local" && request.link) {
-      await capabilityConsent?.onBeforePluginArtifactCommit({
-        pluginId: installed.pluginId,
-        stagedArtifactDir: request.path,
-        mode: request.mode ?? "install",
-      });
-    }
-    const transaction = resolvePluginInstallTransaction(installed);
-    if (completed.expectedPluginId && installed.pluginId !== completed.expectedPluginId) {
-      await transaction?.rollback();
-      return {
-        ok: false as const,
-        error: `official catalog plugin id mismatch: expected ${completed.expectedPluginId}, got ${installed.pluginId}`,
-      };
-    }
-    const persisted = await persistManagedSourceInstall({
-      ...params,
-      snapshot: completed.snapshot ?? params.snapshot,
-      pluginId: installed.pluginId,
-      install: capabilityConsent
-        ? capabilityConsent.applyAcceptedSurface(installed.pluginId, completed.install(installed))
-        : completed.install(installed),
-      transaction,
-      successMessage: completed.successMessage,
-      beforePersistentApply: params.beforePersistentApply,
-    });
-    return {
-      ...installed,
-      config: persisted.config,
-      ...(persisted.warnings.length > 0 ? { warnings: [...new Set(persisted.warnings)] } : {}),
-    };
-  };
-
-  if (request.source === "local") {
-    const installPath = request.link ? request.path : undefined;
-    const linkedSnapshot = request.link
-      ? {
-          ...params.snapshot,
-          config: {
-            ...params.snapshot.config,
-            plugins: {
-              ...params.snapshot.config.plugins,
-              load: {
-                ...params.snapshot.config.plugins?.load,
-                paths: uniqueStrings([
-                  ...(params.snapshot.config.plugins?.load?.paths ?? []),
-                  request.path,
-                ]),
-              },
-            },
-          },
-        }
-      : params.snapshot;
-    return await complete(
-      installPluginFromPath({
-        ...common,
-        path: request.path,
-        mode: request.mode,
-        ...(request.link ? { dryRun: true, allowSourceTypeScriptEntries: true } : {}),
-      }),
-      {
-        snapshot: linkedSnapshot,
-        successMessage: request.successMessage,
-        install: (result) => ({
-          source: request.recordSource,
-          sourcePath: request.recordPath ?? request.path,
-          installPath: installPath ?? result.targetDir,
-          version: result.version,
-        }),
-      },
-    );
-  }
-
-  if (request.source === "marketplace") {
-    return await complete(
-      installPluginFromMarketplace({
-        ...common,
-        marketplace: request.marketplace,
-        plugin: request.plugin,
-        mode: request.mode,
-      }),
-      {
-        install: (result) => ({
-          source: "marketplace",
-          installPath: result.targetDir,
-          version: result.version,
-          marketplaceName: result.marketplaceName,
-          marketplaceSource: result.marketplaceSource,
-          marketplacePlugin: result.marketplacePlugin,
-        }),
-      },
-    );
-  }
-
-  if (request.source === "npm-pack") {
-    return await complete(
-      installPluginFromNpmPackArchive({
-        ...common,
-        archivePath: request.archivePath,
-        mode: request.mode,
-      }),
-      {
-        install: (result) => ({
-          source: "npm",
-          spec: result.npmResolution?.resolvedSpec ?? result.manifestName ?? result.pluginId,
-          sourcePath: request.archivePath,
-          installPath: result.targetDir,
-          ...(result.version ? { version: result.version } : {}),
-          ...buildNpmResolutionFields(result.npmResolution),
-          artifactKind: "npm-pack",
-          artifactFormat: "tgz",
-          ...(result.npmResolution?.integrity
-            ? { npmIntegrity: result.npmResolution.integrity }
-            : {}),
-          ...(result.npmResolution?.shasum ? { npmShasum: result.npmResolution.shasum } : {}),
-          ...(result.npmTarballName ? { npmTarballName: result.npmTarballName } : {}),
-        }),
-      },
-    );
-  }
-
-  if (request.source === "git") {
-    return await complete(
-      installPluginFromGitSpec({ ...common, spec: request.spec, mode: request.mode }),
-      {
-        install: (result) => ({
-          source: "git",
-          spec: request.spec,
-          installPath: result.targetDir,
-          version: result.version,
-          resolvedAt: result.git.resolvedAt,
-          gitUrl: result.git.url,
-          gitRef: result.git.ref,
-          gitCommit: result.git.commit,
-        }),
-      },
-    );
-  }
-
-  if (request.source === "clawhub") {
-    return await complete(
-      installPluginFromClawHub({
-        ...common,
-        spec: request.spec,
-        mode: request.mode,
-        ...(request.expectedPluginId ? { expectedPluginId: request.expectedPluginId } : {}),
-        ...(request.expectedIntegrity ? { expectedIntegrity: request.expectedIntegrity } : {}),
-        ...(request.confirmInstall ? { confirmInstall: request.confirmInstall } : {}),
-      }),
-      {
-        expectedPluginId: request.expectedPluginId,
-        install: (result) => ({
-          ...buildClawHubPluginInstallRecordFields(result.clawhub),
-          spec: request.recordSpec ?? request.spec,
-          installPath: result.targetDir,
-        }),
-      },
-    );
-  }
-
-  const expectedPluginId =
-    request.source === "official" ? request.pluginId : request.expectedPluginId;
-  return await complete(
-    installPluginFromNpmSpec({
-      ...common,
-      spec: request.spec,
-      mode: request.mode,
-      ...(request.source === "official" || request.trustedSourceLinkedOfficialInstall
-        ? { trustedSourceLinkedOfficialInstall: true }
-        : {}),
-      ...(expectedPluginId ? { expectedPluginId } : {}),
-      ...(request.expectedIntegrity ? { expectedIntegrity: request.expectedIntegrity } : {}),
-    }),
-    {
-      expectedPluginId,
-      install: (result) => ({
-        source: "npm",
-        spec: request.pin
-          ? (result.npmResolution?.resolvedSpec ?? request.spec)
-          : (request.recordSpec ?? request.spec),
-        installPath: result.targetDir,
-        ...(result.version ? { version: result.version } : {}),
-        ...buildNpmResolutionFields(result.npmResolution),
-      }),
-    },
-  );
-}
-
-function resolveManagedClawHubInstallRequest(params: {
-  request: Extract<ManagedPluginInstallRequest, { source: "clawhub" }>;
-  officialEntries: readonly OfficialExternalPluginCatalogEntry[];
-  expectedIntegrity?: string;
-}): Extract<ManagedPluginSourceInstallRequest, { source: "clawhub" }> {
-  const packageName = params.request.packageName.trim();
-  const official = resolveOfficialEntryByClawHubPackage(params.officialEntries, packageName);
-  // Pin the runtime id only when the catalog entry declares one; the entry-id
-  // fallback is just the package name and would reject legitimate installs.
-  const expectedPluginId = official ? resolveDeclaredOfficialPluginId(official) : undefined;
-  const hostedOfficial = resolveHostedOfficialEntryByClawHubPackage(
-    params.officialEntries,
-    packageName,
-  );
-  const hostedSource = hostedOfficial
-    ? resolveOfficialExternalPluginInstallSources(hostedOfficial).find(
-        (source) => source.source === "clawhub",
-      )
-    : undefined;
-  const hostedClawHub = parseClawHubPluginSpec(hostedSource?.spec ?? "");
-  const requestMatchesHostedCandidate =
-    !params.request.version || params.request.version === hostedClawHub?.version;
-  const version =
-    params.request.version ?? (requestMatchesHostedCandidate ? hostedClawHub?.version : undefined);
-  const expectedIntegrity =
-    params.expectedIntegrity ??
-    (requestMatchesHostedCandidate ? hostedSource?.expectedIntegrity : undefined);
-  return {
-    source: "clawhub",
-    spec: buildClawHubSpec(packageName, version),
-    ...(official ? { trustedSourceLinkedOfficialInstall: true } : {}),
-    ...(expectedPluginId ? { expectedPluginId } : {}),
-    ...(expectedIntegrity ? { expectedIntegrity } : {}),
-  };
-}
-
-function resolveManagedOfficialInstallRequest(params: {
-  request: Extract<ManagedPluginInstallRequest, { source: "official" }>;
-  officialEntries: readonly OfficialExternalPluginCatalogEntry[];
-}): ManagedPluginSourceInstallRequest {
-  const entry = resolveOfficialEntryById(params.officialEntries, params.request.pluginId);
-  if (!entry) {
-    throw new ManagedPluginLifecycleError(
-      `unknown official plugin catalog entry: ${params.request.pluginId}`,
-    );
-  }
-  const pluginId = resolveOfficialExternalPluginId(entry);
-  const install = resolveOfficialExternalPluginInstall(entry);
-  if (!pluginId || !install) {
-    throw new ManagedPluginLifecycleError(
-      `official plugin catalog entry is not installable: ${params.request.pluginId}`,
-    );
-  }
-  const installSources = resolveOfficialExternalPluginInstallSources(entry);
-  const primary = installSources[0];
-  if (!primary) {
-    throw new ManagedPluginLifecycleError(
-      `official plugin catalog entry has no supported install source: ${params.request.pluginId}`,
-    );
-  }
-  return {
-    source: "official",
-    spec: primary.spec,
-    installSources,
-    pluginId,
-    expectedPluginId: resolveDeclaredOfficialPluginId(entry),
-    mode: "install",
-  };
-}
-
-/** Install a ClawHub or curated official plugin through the canonical install pipeline. */
-export async function installManagedPlugin(params: {
-  request: ManagedPluginInstallRequest;
-  env?: NodeJS.ProcessEnv;
-}): Promise<{ plugin: ManagedPluginCatalogEntry; warnings?: string[] }> {
-  const env = params.env ?? process.env;
-  return await withPluginLifecycleLease({ env }, async () => {
-    const snapshot = await readPluginMutationSnapshot(env);
-    const officialCatalog = await loadOfficialCatalog();
-    const warnings: string[] = [];
-    const installLogger = createInstallLogger(warnings);
-    const request =
-      params.request.source === "clawhub"
-        ? resolveManagedClawHubInstallRequest({
-            request: params.request,
-            officialEntries: officialCatalog.entries,
-          })
-        : resolveManagedOfficialInstallRequest({
-            request: params.request,
-            officialEntries: officialCatalog.entries,
-          });
-    const installed = await installManagedPluginSource({
-      request,
-      snapshot,
-      env,
-      logger: installLogger,
-      ...(params.request.acknowledgeCapabilities
-        ? { acknowledgeCapabilities: params.request.acknowledgeCapabilities }
-        : {}),
-      ...(params.request.acknowledgeInstallPolicyWarning
-        ? {
-            safetyOverrides: {
-              onInstallPolicyWarning: async () => ({ status: "approved" as const }),
-            },
-          }
-        : {}),
-      invalidateRuntimeCache: false,
-      runtime: createSilentRuntime(),
-    });
-    if (!installed.ok) {
-      return throwInstallFailure(installed);
-    }
-    warnings.push(...(installed.warnings ?? []));
-    const workspace = resolvePluginControlPlaneWorkspace({ config: installed.config, env });
-    if (workspace.diagnostic && !getProcessGatewayPluginMetadataSnapshot()) {
-      warnings.push(workspace.diagnostic.message);
-    }
-    // Management inspects the committed candidate; the Gateway keeps its boot inventory.
-    const installedMetadata = refreshManagedPluginMetadata({ config: installed.config, env });
-    const catalog = await listManagedPlugins({
-      config: installed.config,
-      env,
-      officialCatalog,
-      metadata: installedMetadata,
-    });
-    const installedOwnership = createInstalledPluginOwnershipResolver(
-      installedMetadata.index,
-      env,
-    ).resolvePackage(installed.pluginId);
-    if (!installedOwnership.ok) {
-      throw new ManagedPluginLifecycleError(installedOwnership.error);
-    }
-    const installedPluginIds = installedOwnership.value.pluginIds;
-    const representativePluginId = installedPluginIds[0]!;
-    const plugin = catalog.plugins.find((entry) => entry.id === representativePluginId);
-    if (!plugin) {
-      throw new ManagedPluginLifecycleError(
-        `installed plugin missing from refreshed registry: ${installed.pluginId}`,
-      );
-    }
-    return {
-      plugin,
-      ...(installedPluginIds.length > 1 || warnings.length > 0
-        ? {
-            warnings: [
-              ...(installedPluginIds.length > 1
-                ? [
-                    `Installed package "${installed.pluginId}" with plugin entries: ${installedPluginIds.join(", ")}.`,
-                  ]
-                : []),
-              ...new Set(warnings),
-            ],
-          }
-        : {}),
-    };
-  });
-}
-
-/** Persist desired plugin policy while preserving allow/deny, slot, include, and hash guards. */
-export async function setManagedPluginEnabled(params: {
-  pluginId: string;
-  enabled: boolean;
-  acknowledgeCapabilities?: PluginCapabilityConsentAcknowledgment;
-  env?: NodeJS.ProcessEnv;
-}): Promise<{
-  plugin: ManagedPluginCatalogEntry;
-  changedPaths: string[];
-  warnings?: string[];
-}> {
-  const env = params.env ?? process.env;
-  return await withPluginLifecycleLease({ env }, async () => {
-    const snapshot = await readPluginMutationSnapshot(env);
-    const metadata = loadFreshManagedPluginMetadata(snapshot.config, env);
-    const pluginId = metadata.normalizePluginId(params.pluginId.trim());
-    const installedPlugin = metadata.index.plugins.find((plugin) => plugin.pluginId === pluginId);
-    if (!installedPlugin) {
-      throw new ManagedPluginLifecycleError(`plugin not installed: ${params.pluginId}`);
-    }
-    if (params.enabled && !installedPlugin.enabled) {
-      await resolvePluginCapabilityConsent({
-        config: snapshot.config,
-        env,
-        pluginId,
-        acknowledge: params.acknowledgeCapabilities,
-        metadata,
-      });
-    }
-    let next = snapshot.config;
-    const warnings: string[] = [];
-    let policyPluginId = pluginId;
-    if (params.enabled) {
-      // The admin-scoped enable RPC is an explicit trust action. Preserve the
-      // existing inventory while admitting only the selected installed plugin.
-      if ((next.plugins?.allow?.length ?? 0) > 0) {
-        next = ensurePluginAllowlisted(next, pluginId);
-      }
-      const enableResult = enableExplicitlySelectedPluginInConfig(next, pluginId, {
-        updateChannelConfig: false,
-      });
-      if (!enableResult.enabled) {
-        throw new ManagedPluginLifecycleError(
-          `plugin "${pluginId}" could not be enabled (${enableResult.reason ?? "unknown reason"})`,
-        );
-      }
-      next = enableResult.config;
-      policyPluginId = enableResult.pluginId;
-      const slotResult = applySlotSelectionForPlugin(next, pluginId, metadata);
-      next = slotResult.config;
-      warnings.push(...slotResult.warnings);
-    } else {
-      next = setPluginEnabledInConfig(next, pluginId, false, { updateChannelConfig: false });
-    }
-    const changedPaths = new Set<string>();
-    collectChangedPaths(snapshot.config, next, "", changedPaths);
-    await replaceConfigFile({
-      nextConfig: next,
-      baseHash: snapshot.baseHash,
-      writeOptions: snapshot.writeOptions,
-    });
-    await refreshPluginRegistryAfterConfigMutation({
-      config: next,
-      env,
-      reason: "policy-changed",
-      invalidateRuntimeCache: false,
-      policyPluginIds: [policyPluginId],
-      logger: { warn: (message) => warnings.push(message) },
-    });
-    const updatedMetadata = refreshManagedPluginMetadata({ config: next, env });
-    const catalog = await listManagedPlugins({ config: next, env, metadata: updatedMetadata });
-    const plugin = catalog.plugins.find((entry) => entry.id === pluginId);
-    if (!plugin) {
-      throw new ManagedPluginLifecycleError(
-        `updated plugin missing from refreshed registry: ${pluginId}`,
-      );
-    }
-    return {
-      plugin,
-      changedPaths: [...changedPaths].filter(Boolean).toSorted(),
-      ...(warnings.length > 0 ? { warnings } : {}),
-    };
-  });
-}
-
-/** Remove an installed plugin: config references, install record, and managed files. */
-export async function uninstallManagedPlugin(params: {
-  pluginId: string;
-  env?: NodeJS.ProcessEnv;
-}): Promise<{ pluginId: string; removed: string[]; warnings?: string[] }> {
-  const env = params.env ?? process.env;
-  return await withPluginLifecycleLease({ env }, async () => {
-    const snapshot = await readPluginMutationSnapshot(env);
-    const installRecords = await loadInstalledPluginIndexInstallRecords({ env });
-    // Mirror the CLI uninstall flow: plan against config carrying install records
-    // so managed npm/git directories resolve, then persist the stripped config.
-    const configWithRecords = withPluginInstallRecords(snapshot.config, installRecords);
-    const metadata = loadFreshManagedPluginMetadata(configWithRecords, env);
-    const pluginId = metadata.normalizePluginId(params.pluginId.trim());
-    const record = metadata.index.plugins.find((plugin) => plugin.pluginId === pluginId);
-    if (record?.origin === "bundled") {
-      throw new ManagedPluginLifecycleError(
-        `bundled plugin cannot be uninstalled: ${pluginId}; disable it instead`,
-      );
-    }
-    if (!record && !Object.hasOwn(installRecords, pluginId)) {
-      throw new ManagedPluginLifecycleError(`Plugin not found: ${pluginId}`);
-    }
-    const ownership = createInstalledPluginOwnershipResolver(metadata.index, env).resolveLifecycle(
-      pluginId,
-    );
-    if (!ownership.ok) {
-      throw new ManagedPluginLifecycleError(ownership.error);
-    }
-    const { installOwner, pluginIds: ownedPluginIds } = ownership.value;
-    const policyPluginIds = ownedPluginIds.length > 0 ? ownedPluginIds : [installOwner];
-    const ownedManifests = ownedPluginIds.flatMap((entryId) => {
-      const manifest = metadata.byPluginId.get(entryId);
-      return manifest ? [manifest] : [];
-    });
-    const channelIds =
-      ownedManifests.length > 0
-        ? uniqueStrings(ownedManifests.flatMap((manifest) => manifest.channels))
-        : ownership.value.kind === "orphan" &&
-            createInstalledPluginIndexScopeLookup(metadata.index).hasChannelContributionOwners([
-              installOwner,
-            ])
-          ? []
-          : undefined;
-    const extensionsDir = resolveDefaultPluginExtensionsDir(env);
-    const initialPlan = planPluginUninstall(
-      recordPluginPackageUninstallPlan(
-        {
-          config: configWithRecords,
-          pluginId: installOwner,
-          ...(channelIds !== undefined ? { channelIds } : {}),
-          deleteFiles: true,
-          extensionsDir,
-        },
-        {
-          runtimePluginIds: policyPluginIds,
-          runtimeLoadPaths: ownedPluginIds.flatMap(
-            (entryId) => metadata.byPluginId.get(entryId)?.source ?? [],
-          ),
-        },
-      ),
-    );
-    if (!initialPlan.ok) {
-      throw new ManagedPluginLifecycleError(initialPlan.error);
-    }
-    let plan = initialPlan;
-    let finalSnapshot = snapshot;
-    let directoryResult = { directoryRemoved: false, warnings: [] as string[] };
-    if (plan.directoryRemoval) {
-      const disabledConfig = prepareConfigForDisabledPluginSet(snapshot.config, policyPluginIds);
-      await replaceConfigFile({
-        nextConfig: disabledConfig,
-        baseHash: snapshot.baseHash,
-        writeOptions: {
-          ...snapshot.writeOptions,
-          afterWrite: { mode: "auto" },
-        },
-      });
-      directoryResult = await applyPluginUninstallDirectoryRemoval(plan.directoryRemoval);
-      if (pluginUninstallTargetExists(plan.directoryRemoval.target)) {
-        throw new ManagedPluginLifecycleError(
-          `Failed to remove plugin directory ${plan.directoryRemoval.target}; the plugin remains disabled and tracked so uninstall can be retried.`,
-          { kind: "unavailable" },
-        );
-      }
-      finalSnapshot = await readPluginMutationSnapshot(env);
-      const refreshedConfigWithRecords = withPluginInstallRecords(
-        finalSnapshot.config,
-        installRecords,
-      );
-      const refreshedPlan = planPluginUninstall(
-        recordPluginPackageUninstallPlan(
-          {
-            config: refreshedConfigWithRecords,
-            pluginId: installOwner,
-            ...(channelIds !== undefined ? { channelIds } : {}),
-            deleteFiles: true,
-            extensionsDir,
-          },
-          {
-            runtimePluginIds: policyPluginIds,
-            runtimeLoadPaths: ownedPluginIds.flatMap(
-              (entryId) => metadata.byPluginId.get(entryId)?.source ?? [],
-            ),
-          },
-        ),
-      );
-      if (!refreshedPlan.ok) {
-        throw new ManagedPluginLifecycleError(refreshedPlan.error);
-      }
-      plan = refreshedPlan;
-    }
-    const nextConfig = withoutPluginInstallRecords(plan.config);
-    const nextInstallRecords = removePluginInstallRecordFromRecords(installRecords, installOwner);
-    await commitPluginInstallRecordsWithConfig({
-      previousInstallRecords: installRecords,
-      nextInstallRecords,
-      nextConfig,
-      baseHash: finalSnapshot.baseHash,
-      writeOptions: finalSnapshot.writeOptions,
-    });
-    const warnings = [
-      ...collectClawPluginUninstallWarnings({
-        pluginId: installOwner,
-        installRecord: installRecords[installOwner],
-        env,
-      }),
-      ...(pluginId !== installOwner || ownedPluginIds.length > 1
-        ? [
-            `Uninstalled package "${installOwner}" and all owned plugin entries: ${ownedPluginIds.join(", ")}.`,
-          ]
-        : []),
-      ...directoryResult.warnings,
-    ];
-    await refreshPluginRegistryAfterConfigMutation({
-      config: nextConfig,
-      env,
-      reason: "source-changed",
-      installRecords: nextInstallRecords,
-      invalidateRuntimeCache: false,
-      logger: { warn: (message) => warnings.push(message) },
-    });
-    refreshManagedPluginMetadata({ config: nextConfig, env });
-    const removed = formatUninstallActionLabels({
-      ...plan.actions,
-      directory: directoryResult.directoryRemoved,
-    });
-    return {
-      pluginId: installOwner,
-      removed,
-      ...(warnings.length > 0 ? { warnings: [...new Set(warnings)] } : {}),
-    };
-  });
-}
-
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
