+++ src/system-agent/plugin-artifact.ts
@@ -285,7 +285,7 @@ export async function executePluginArtifactActivation(
         return await withVerifiedArtifact(bytes, async ({ archivePath, review }) => {
           const { assertConfigWriteAllowedInCurrentMode } = await import("../config/config.js");
           const { loadConfigForInstall } = await import("../cli/plugins-install-config.js");
-          const { installManagedPluginSource } = await import("../plugins/management-service.js");
+          const { installManagedPluginSource } = await import("../plugins/management-install.js");
           assertConfigWriteAllowedInCurrentMode();
           const assertPersistentApply = () => {
             ctx.assertPersistentApply?.();
